Moved vertical tab flyout to title bar
