+++ Files/Strings/hu-HU/Resources.resw
@@ -2169,9 +2169,6 @@
   <data name="NavToolbarWidgets.Label" xml:space="preserve">
     <value>Widgetek</value>
   </data>
-  <data name="VerticalTabStrip.Label" xml:space="preserve">
-    <value>Lapok</value>
-  </data>
   <data name="NavToolbarSelectionOptions.Label" xml:space="preserve">
     <value>Kiválasztás</value>
   </data>
@@ -2279,5 +2276,35 @@
   </data>
   <data name="SecurityUnknownOwnerText.Text" xml:space="preserve">
     <value>Ismeretlen tulajdonos</value>
+  </data>
+  <data name="SettingsNavSidebar.Content" xml:space="preserve">
+    <value>Oldalsáv</value>
+  </data>
+  <data name="SettingsSidebarTitle.Text" xml:space="preserve">
+    <value>Oldalsáv</value>
+  </data>
+  <data name="SettingsVerticalTabFlyoutSwitch.Header" xml:space="preserve">
+    <value>Lapok megjelenítése a címsoron</value>
+  </data>
+  <data name="DefaultSkin" xml:space="preserve">
+    <value>Alapértelmezett</value>
+  </data>
+  <data name="SettingsAppearanceCustomSkins.Header" xml:space="preserve">
+    <value>Alternatív kinézet</value>
+  </data>
+  <data name="SettingsAppearanceOpenSkinsFolderButton.Content" xml:space="preserve">
+    <value>Kinézetek mappa megnyitása</value>
+  </data>
+  <data name="SettingsSkinsLearnMoreButton.AutomationProperties.Name" xml:space="preserve">
+    <value>Tudjon meg többet az alternatív kinézetekről</value>
+  </data>
+  <data name="SettingsSkinsLearnMoreButton.ToolTipService.ToolTip" xml:space="preserve">
+    <value>Tudjon meg többet az alternatív kinézetekről</value>
+  </data>
+  <data name="SettingsSkinsTeachingTipHeader.Text" xml:space="preserve">
+    <value>Az alternatív kinézetek remek lehetőséget nyújtanak a Files személyre szabására.</value>
+  </data>
+  <data name="SettingsSkinsTeachingTipHyperlinkText.Text" xml:space="preserve">
+    <value>Dokumentáció megtekintése.</value>
   </data>
 </root>
