+++ Files/Strings/id-ID/Resources.resw
@@ -1203,9 +1203,6 @@
   <data name="PropertyAddress" xml:space="preserve">
     <value>Alamat</value>
   </data>
-  <data name="VerticalTabStrip.Label" xml:space="preserve">
-    <value>Tab</value>
-  </data>
   <data name="NavToolbarSelectionOptions.AutomationProperties.Name" xml:space="preserve">
     <value>Opsi pilihan</value>
   </data>
