+++ Files/Views/MainPage.xaml
@@ -252,7 +252,8 @@
                 Canvas.ZIndex="4"
                 IsHitTestVisible="True">
                 <Grid.ColumnDefinitions>
-                    <ColumnDefinition />
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
                     <ColumnDefinition x:Name="RightPaddingColumn" Width="0" />
                 </Grid.ColumnDefinitions>
                 <usercontrols:HorizontalMultitaskingControl
@@ -263,6 +264,47 @@
                     x:FieldModifier="public"
                     Canvas.ZIndex="40"
                     Loaded="HorizontalMultitaskingControl_Loaded" />
+                <Button
+                    x:Name="VerticalTabStripInvokeButton"
+                    x:Uid="VerticalTabStrip"
+                    Grid.Column="1"
+                    Height="28"
+                    Padding="8,4,8,4"
+                    HorizontalAlignment="Right"
+                    x:Load="{x:Bind AppSettings.IsVerticalTabFlyoutEnabled, Mode=OneWay}"
+                    AccessKey="T"
+                    AllowDrop="True"
+                    AutomationProperties.Name="Vertical tab strip"
+                    Background="Transparent"
+                    BorderBrush="Transparent"
+                    BorderThickness="0"
+                    DragEnter="VerticalTabStripInvokeButton_DragEnter"
+                    IsTapEnabled="True"
+                    Loaded="VerticalTabStripInvokeButton_Loaded"
+                    PointerEntered="Button_PointerEntered"
+                    PointerExited="VerticalTabStripInvokeButton_PointerExited"
+                    Tapped="VerticalTabStrip_Tapped"
+                    ToolTipService.ToolTip="Open tabs">
+                    <Button.Content>
+                        <FontIcon
+                            FontSize="12"
+                            Foreground="{ThemeResource TextControlPlaceholderForegroundFocused}"
+                            Glyph="&#xF5ED;" />
+                    </Button.Content>
+                    <Button.Flyout>
+                        <Flyout
+                            x:Name="VerticalTabViewFlyout"
+                            Closed="VerticalTabViewFlyout_Closed"
+                            Opened="VerticalTabViewFlyout_Opened">
+                            <usercontrols:VerticalTabViewControl
+                                x:Name="VerticalTabs"
+                                Width="275"
+                                MaxHeight="300"
+                                x:FieldModifier="public"
+                                PointerExited="VerticalTabs_PointerExited" />
+                        </Flyout>
+                    </Button.Flyout>
+                </Button>
             </Grid>
         </controls:SidebarControl.TabContent>
     </controls:SidebarControl>
